Stabilize Utils tests for JVM and GPX builder behavior
